requirements.txt
pandas==2.0.0
numpy==1.24.3
matplotlib==3.7.1
seaborn==0.12.2
plotly==5.14.1
dash==2.9.3
streamlit==1.22.0
scikit-learn==1.2.2
jupyter==1.0.0

# 数据库连接
pymysql==1.1.0
psycopg2-binary==2.9.6
sqlalchemy==2.0.8
sqlite3==0.0.0

# API请求
requests==2.28.2

# 数据处理增强
openpyxl==3.1.2
xlrd==2.0.1
statsmodels==0.14.0 